Mejor División del Tablero Todo está más organizado y repartido para que tenga la misma lógica que el resto de directorios.
